hwbench/stressng: Split Stream specific code
As per issue #48, let's split stress-ng specific code into different
files to increase readability.
